requirements.txt
fastapi
uvicorn
pdfplumber
sentence-transformers
faiss-cpu
easyocr
Pillow
clip 
transformers
torch
typing_extensions
jinja2
starlette
python-multipart
rank_bm25
pymupdf 
